+++ src/test/java/Runners/TestRunner.java
@@ -1,4 +1,4 @@
-package Tests;
+package Runners;
 
 import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
