+++ src/test/java/StepDefinitions/NewsletterDefs.java
@@ -1,6 +1,6 @@
 package StepDefinitions;
 
-import Tests.DriverFactory;
+import Util.DriverFactory;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
