+++ src/test/java/StepDefinitions/SignInDefs.java
@@ -1,8 +1,7 @@
 package StepDefinitions;
 
-import Tests.DriverFactory;
+import Util.DriverFactory;
 import io.cucumber.datatable.DataTable;
-import io.cucumber.java.an.E;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
